1: tweak ECF to not show `v` for the address
the address class has a superclass of method, which has fields including
`v` for value. in the address instance though, we don't want to show
this value as we're using `al` (address lines) instead, so the new
`ignore` line here sets this up so that the record editor won't output
this class within the `a` (address) class.